requirements.txt
numpy
pandas
scikit-learn
requests
beautifulsoup4
python-dotenv
langchain==0.1.9
serpapi==0.1.5
einops
accelerate
bitsandbytes
faknow==0.0.3
streamlit==1.32.2
gdown==4.6.0
joblib~=1.2.0
gensim~=4.3.0
nltk~=3.8.1
openai==0.28
rouge-score
langchain_community
openpyxl

